:recycle: refactor: 서비스 클래스 구조 개선 (#62)
